Small changes in detail
1.NitroLz.cpp Line 277 :: Usage of "++=" has been standardized into "="
and "++"
2.class initialization reordered
3.change "std::memcpy" to "memcpy"
4.other small fixes

+++ src/KssuFile.cpp
@@ -369,7 +369,7 @@ void KfMap::readFile(const u8* src){
         std::memcpy(pScriptList->back().data(),p,len);
         p+=len;
     }
-    assert(p-rawScripts.get()==rawScriptsLength-1);
+    assert((u32)(p-rawScripts.get())==(rawScriptsLength-1));
     loaded=true;
